test(image_transport_decompressor): move the matrix to a logic test
* move the 40-case encoding/format matrix out of the node test into a
  new test_image_transport_decompressor.cpp that calls decompress()
  directly, so it needs no ROS context
* the node test keeps only the 3 wiring cases: publish, "encoding"
  parameter forwarding, and drop-without-publish on a decode failure
* register the new logic test in CMakeLists.txt

The expected values are carried over unchanged, so the move itself
proves the relocation is faithful.
