+++ sensing/autoware_image_transport_decompressor/CMakeLists.txt
@@ -33,6 +33,13 @@ rclcpp_components_register_node(${PROJECT_NAME}_ros
 if(BUILD_TESTING)
   find_package(ament_cmake_ros REQUIRED)
 
+  ament_add_gtest(test_image_transport_decompressor
+    test/test_image_transport_decompressor.cpp
+  )
+  target_link_libraries(test_image_transport_decompressor
+    ${PROJECT_NAME}
+  )
+
   ament_add_ros_isolated_gtest(test_image_transport_decompressor_node
     test/test_image_transport_decompressor_node.cpp
   )
